+++ src/utils.ts
@@ -1337,7 +1337,7 @@ export class newShape {
     throw new Error(`Method computeTooltipOrigin not implemented for ${this.constructor.name}`)
   }
 
-  public draw(context: CanvasRenderingContext2D) {
+  public draw(context: CanvasRenderingContext2D): void {
     const scaledPath = new Path2D();
     const contextMatrix = context.getTransform();
     scaledPath.addPath(this.path, new DOMMatrix().scale(contextMatrix.a, contextMatrix.d));
@@ -1352,9 +1352,11 @@ export class newShape {
     context.restore();
   }
 
+  public drawTooltip(plotOrigin: Vertex, plotSize: Vertex, context: CanvasRenderingContext2D): void {}
+
   public mouseDown(canvasMouse: Vertex, frameMouse: Vertex) { }
 
-  public mouseMove(canvasMouse: Vertex, frameMouse: Vertex) { return false }
+  public mouseMove(canvasMouse: Vertex, frameMouse: Vertex): boolean { return false }
 
   public mouseUp() { }
 }
@@ -1661,7 +1663,7 @@ export class newText extends newShape {
     return `${style} ${fontsize}px ${font}`
   }
 
-  get fullFont() { return newText.buildFont(this.style, this.fontsize, this.font) }
+  get fullFont(): string { return newText.buildFont(this.style, this.fontsize, this.font) }
 
   private automaticFontSize(context: CanvasRenderingContext2D): number {
     let tmp_context: CanvasRenderingContext2D = context;
@@ -1708,7 +1710,7 @@ export class newText extends newShape {
 
   public capitalizeSelf(): void { this.text = newText.capitalize(this.text) }
 
-  public draw(context: CanvasRenderingContext2D) {
+  public draw(context: CanvasRenderingContext2D): void {
     context.save();
     const writtenText = this.format(context);
     context.font = this.fullFont;
@@ -1752,7 +1754,7 @@ export class newText extends newShape {
     return writtenText
   }
 
-  private write(writtenText: string[], context: CanvasRenderingContext2D) {
+  private write(writtenText: string[], context: CanvasRenderingContext2D): void {
     context.fillStyle = this.fillStyle
     if (writtenText.length != 1) {
       var offset: number = writtenText.length - 1;
@@ -1762,7 +1764,7 @@ export class newText extends newShape {
     }
   }
 
-  public removeEndZeros() {
+  public removeEndZeros(): void {
     let splitText = this.text.split(".");
     if (splitText.length > 1) {
       let splitDecimal = splitText[1].split("e");
@@ -1776,7 +1778,7 @@ export class newText extends newShape {
     }
   }
 
-  private cutting_text(context: CanvasRenderingContext2D, maxWidth: number) {
+  private cutting_text(context: CanvasRenderingContext2D, maxWidth: number): string[] {
     var words = this.text.split(' ');
     var space_length = context.measureText(' ').width;
     var cut_texts = [];
@@ -1837,7 +1839,7 @@ export class newPoint2D extends Vertex {
     this.path = this.buildPath();
   };
 
-  get drawnShape() {
+  get drawnShape(): newShape {
     let marker = new newShape();
     if (CIRCLES.indexOf(this.marker) > -1) { marker = new newCircle(this.coordinates, this.size) }
     if (MARKERS.indexOf(this.marker) > -1) { marker = new Mark(this.coordinates, this.size) };
@@ -1871,7 +1873,7 @@ export class newPoint2D extends Vertex {
 
   private buildPath(): Path2D { return this.drawnShape.path };
 
-  public draw(context: CanvasRenderingContext2D) {
+  public draw(context: CanvasRenderingContext2D): void {
     this.path = this.buildPath();
     context.lineWidth = this.lineWidth;
     context.fillStyle = this.color;
@@ -1908,26 +1910,26 @@ export class Bar extends newRect {
     return new Vertex(this.origin.x + this.size.x / 2, this.origin.y + this.size.y).transform(contextMatrix)
   }
 
-  public setGeometry(origin: Vertex, size: Vertex) {
+  public setGeometry(origin: Vertex, size: Vertex): void {
     this.origin = origin;
     this.size = size;
   }
 
-  public draw(context: CanvasRenderingContext2D) {
+  public draw(context: CanvasRenderingContext2D): void {
     if (this.size.x != 0 && this.size.y != 0) {
       super.draw(context);
       this.tooltipOrigin = this.computeTooltipOrigin(context.getTransform());
     }
   }
 
-  public drawTooltip(context: CanvasRenderingContext2D) {
+  public drawTooltip(plotOrigin: Vertex, plotSize: Vertex, context: CanvasRenderingContext2D): void {
     if (this.isClicked) {
       const tooltip = new newTooltip(this.tooltipOrigin, this.tooltipMap, context);
-      tooltip.draw(context);
+      tooltip.draw(plotOrigin, plotSize, context);
     }
   }
 
-  public computeStats(values: number[], precision: number) {
+  public computeStats(values: number[], precision: number): void {
     this.min = Math.round(Math.min(...values) * precision) / precision;
     this.max = Math.round(Math.max(...values) * precision) / precision;
     this.mean = Math.round(values.reduce((a, b) => a + b, 0) / values.length * precision) / precision;
@@ -1958,6 +1960,8 @@ export class newTooltip {
     }
 
   private buildText(context: CanvasRenderingContext2D): [string[], Vertex] {
+    context.save();
+    context.font = `${this.fontsize}px sans-serif`
     let printedRows = ['Information: '];
     let textLength = context.measureText(printedRows[0]).width;
     this.dataToPrint.forEach((value, key) => {
@@ -1966,6 +1970,7 @@ export class newTooltip {
       if (textWidth > textLength) { textLength = textWidth };
       printedRows.push(text);
     })
+    context.restore();
     return [printedRows, new Vertex(textLength + TOOLTIP_TEXT_OFFSET * 2, (printedRows.length + 1.5) * this.fontsize)]
   }
 
@@ -1986,7 +1991,7 @@ export class newTooltip {
     return textOrigin.add(new Vertex(textOffsetX, textOffsetY));
   }
 
-  private writeText(textOrigin: Vertex, context: CanvasRenderingContext2D) {
+  private writeText(textOrigin: Vertex, context: CanvasRenderingContext2D): void {
     this.printedRows.forEach((row, index) => {
       textOrigin.y += index == 0 ? 0 : this.fontsize;
       const text = new newText(row, textOrigin, {fontsize: this.fontsize, baseline: "middle", style: index == 0 ? 'bold' : ''});
@@ -1995,44 +2000,45 @@ export class newTooltip {
     })
   }
 
-  private insideCanvas(canvasSize: Vertex, scaling: Vertex): boolean {
+  private insideCanvas(plotOrigin: Vertex, plotSize: Vertex, scaling: Vertex): boolean {
     let isInside = true;
     const downLeftCorner = this.squareOrigin.add(new Vertex(-this.size.x / 2, TOOLTIP_TRIANGLE_SIZE).scale(scaling));
     const upRightCorner = downLeftCorner.add(this.size.scale(scaling));
-    const upRightDiff = canvasSize.subtract(upRightCorner);
+    const upRightDiff = plotOrigin.add(plotSize).subtract(upRightCorner);
+    const downLeftDiff = downLeftCorner.subtract(plotOrigin);
 
     if (upRightDiff.x < 0) {
       this.squareOrigin.x += upRightDiff.x;
-    } else if (upRightDiff.x > canvasSize.x) {
-      this.squareOrigin.x += upRightDiff.x - canvasSize.x;
+    } else if (upRightDiff.x > plotSize.x) {
+      this.squareOrigin.x += upRightDiff.x - plotSize.x;
     }
     if (upRightDiff.y < 0) {
       this.squareOrigin.y += upRightDiff.y;
       this.origin.y += upRightDiff.y;
-    } else if (upRightDiff.y > canvasSize.y){
-      this.squareOrigin.y += upRightDiff.y - canvasSize.y;
-      this.origin.y += upRightDiff.y - canvasSize.y;
+    } else if (upRightDiff.y > plotSize.y){
+      this.squareOrigin.y += upRightDiff.y - plotSize.y;
+      this.origin.y += upRightDiff.y - plotSize.y;
     }
-// To handle once one case is met
-    // const downLeftDiff = canvasSize.subtract(downLeftCorner);
-    // if (downLeftCorner.y < 0) {
-    //   console.log("down inf y")
-    // } else if (downLeftCorner.y > canvasSize.y){
-    //   console.log("down sup y")
-    // }
-    // if (upRightCorner.y < 0) {
-    //   console.log("down inf y")
-    // } else if (upRightCorner.y > canvasSize.y){
-    //   console.log("down sup y")
-    // }
+
+    if (downLeftDiff.x < 0) {
+      this.squareOrigin.x -= downLeftDiff.x;
+    } else if (downLeftDiff.x > plotSize.x){
+      this.squareOrigin.x -= downLeftDiff.x - plotSize.x;
+    }
+    if (downLeftDiff.y < 0) { // Maybe wrong, did not met the case
+      this.squareOrigin.y -= downLeftDiff.y;
+      this.origin.y -= downLeftDiff.y;
+    } else if (downLeftDiff.y > plotSize.y){
+      this.squareOrigin.y += downLeftDiff.y - plotSize.y;
+      this.origin.y += downLeftDiff.y - plotSize.y;
+    }
     return isInside
   }
 
-  public draw(context: CanvasRenderingContext2D) {
+  public draw(plotOrigin: Vertex, plotSize: Vertex, context: CanvasRenderingContext2D): void {
     const contextMatrix = context.getTransform();
-    const canvasSize = new Vertex(context.canvas.width, context.canvas.height);
     const scaling = new Vertex(1 / contextMatrix.a, 1 / contextMatrix.d);
-    this.insideCanvas(canvasSize, scaling);
+    this.insideCanvas(plotOrigin, plotSize, scaling);
     const textOrigin = this.computeTextOrigin(scaling);
     const scaledPath = new Path2D();
     this.squareOrigin = this.squareOrigin.scale(scaling);
@@ -2052,7 +2058,7 @@ export class newTooltip {
 }
 
 export class newAxis {
-  public ticksCoords: newPoint2D[];
+  public ticksPoints: newPoint2D[];
   public drawPath: Path2D;
   public path: Path2D;
   public lineWidth: number = 1;
@@ -2075,6 +2081,8 @@ export class newAxis {
   private _marginRatio: number = 0.1;
   private _minValue: number;
   private _maxValue: number;
+  private _initMinValue: number;
+  private _initMaxValue: number;
   private _previousMin: number;
   private _previousMax: number;
   private offsetTicks: number;
@@ -2103,7 +2111,7 @@ export class newAxis {
       this.isDiscrete = typeof vector[0] == 'string';
       if (this.isDiscrete) { this.labels = newAxis.uniqueValues(vector) };
       const [minValue, maxValue] = this.computeMinMax(vector);
-      [this._previousMin, this._previousMax] = [this.minValue, this.maxValue] = this.marginedBounds(minValue, maxValue);
+      [this._previousMin, this._previousMax] = [this._initMinValue, this._initMaxValue] = [this.minValue, this.maxValue] = this.marginedBounds(minValue, maxValue);
       this.ticks = this.computeTicks();
       if (!this.isDiscrete) { this.labels = this.numericLabels() };
       this.drawPath = this.buildDrawPath();
@@ -2117,18 +2125,10 @@ export class newAxis {
     return this.isVertical ? Math.abs(this.origin.y - this.end.y) : Math.abs(this.origin.x - this.end.x);
   }
 
-  public computeTextBoxes(context: CanvasRenderingContext2D) {
-    context.save();
-    const calibratedTickText = new newText("88.88e+88", new Vertex(0, 0), { fontsize: this.FONT_SIZE, font: this.FONT });
-    context.font = calibratedTickText.fullFont;
-    const calibratedMeasure = context.measureText(calibratedTickText.text).width;
-    this.maxTickWidth = Math.min(this.freeSpace - this.offsetTicks - 1, calibratedMeasure);
-    this.maxTickHeight = Math.min(this.freeSpace - this.offsetTicks - 1, calibratedTickText.fontsize);
-    if (this.centeredTitle) {
-      const FREE_SPACE = this.freeSpace - this.FONT_SIZE - 0.3 * this.maxTickHeight;
-      this.offsetTitle = (this.isVertical ? Math.min(FREE_SPACE, calibratedMeasure) : Math.min(FREE_SPACE, this.FONT_SIZE * 1.5 + this.offsetTicks));
-    }
-    context.restore();
+  private get drawingColor(): string {
+    let color = this.strokeStyle;
+    if (this.mouseStyleON) { color = this.isHovered ? this.hoverStyle : this.isClicked ? this.clickedStyle : this.strokeStyle };
+    return color
   }
 
   get interval(): number { return Math.abs(this.maxValue - this.minValue) };
@@ -2158,7 +2158,7 @@ export class newAxis {
     return this._nTicks
   }
 
-  get ticks() { return this._ticks }
+  get ticks(): number[] { return this._ticks }
 
   set ticks(value: number[]) { this._ticks = value }
 
@@ -2166,18 +2166,25 @@ export class newAxis {
 
   get transformMatrix(): DOMMatrix { return this.getValueToDrawMatrix() };
 
-  private horizontalPickIdx() { return Math.sign(1 - Math.sign(this.initScale.y)) }
+  private horizontalPickIdx(): number { return Math.sign(1 - Math.sign(this.initScale.y)) }
 
-  private verticalPickIdx() { return Math.sign(1 - Math.sign(this.initScale.x)) }
+  private verticalPickIdx(): number { return Math.sign(1 - Math.sign(this.initScale.x)) }
 
-  public transform(newOrigin: Vertex, newEnd: Vertex) {
+  public transform(newOrigin: Vertex, newEnd: Vertex): void {
     this.origin = newOrigin;
     this.end = newEnd;
     this.drawPath = this.buildDrawPath();
     this.path = this.buildPath();
   }
 
-  public reset(): void { this.rubberBand.reset() }
+  public reset(): void {
+    this.rubberBand.reset();
+    this.minValue = this._initMinValue;
+    this.maxValue = this._initMaxValue;
+    this._previousMin = this._initMinValue;
+    this._previousMax = this._initMaxValue;
+    this.updateTicks();
+  }
 
   private static nearestFive(value: number): number {
     const tenPower = Math.floor(Math.log10(Math.abs(value)));
@@ -2226,6 +2233,20 @@ export class newAxis {
     return [Math.min(...vector), Math.max(...vector)]
   }
 
+  public computeTextBoxes(context: CanvasRenderingContext2D): void {
+    context.save();
+    const calibratedTickText = new newText("88.88e+88", new Vertex(0, 0), { fontsize: this.FONT_SIZE, font: this.FONT });
+    context.font = calibratedTickText.fullFont;
+    const calibratedMeasure = context.measureText(calibratedTickText.text).width;
+    this.maxTickWidth = Math.min(this.freeSpace - this.offsetTicks - 1, calibratedMeasure);
+    this.maxTickHeight = Math.min(this.freeSpace - this.offsetTicks - 1, calibratedTickText.fontsize);
+    if (this.centeredTitle) {
+      const FREE_SPACE = this.freeSpace - this.FONT_SIZE - 0.3 * this.maxTickHeight;
+      this.offsetTitle = (this.isVertical ? Math.min(FREE_SPACE, calibratedMeasure) : Math.min(FREE_SPACE, this.FONT_SIZE * 1.5 + this.offsetTicks));
+    }
+    context.restore();
+  }
+
   private computeTicks(): number[] {
     const increment = newAxis.nearestFive((this.maxValue - this.minValue) / this.nTicks);
     const remainder = this.minValue % increment;
@@ -2236,28 +2257,31 @@ export class newAxis {
     return ticks
   }
 
-  public draw(context: CanvasRenderingContext2D) {
-    context.lineWidth = this.lineWidth;
-    let color = this.strokeStyle;
-    if (this.mouseStyleON) { color = this.isHovered ? this.hoverStyle : this.isClicked ? this.clickedStyle : this.strokeStyle };
+  public draw(context: CanvasRenderingContext2D): void {
+    const canvasHTMatrix = context.getTransform();
+    const pointHTMatrix = canvasHTMatrix.multiply(this.transformMatrix);
+    const color = this.drawingColor;
+
     context.strokeStyle = color;
     context.fillStyle = color;
+    context.lineWidth = this.lineWidth;
     context.stroke(this.drawPath);
     context.fill(this.drawPath);
 
-    const canvasHTMatrix = context.getTransform();
-    const pointHTMatrix = canvasHTMatrix.multiply(this.transformMatrix);
     context.setTransform(pointHTMatrix);
-    this.ticksCoords = this.drawTicks(context, pointHTMatrix, color);
-    context.resetTransform();
+    const [ticksPoints, ticksTexts] = this.drawTicksPoints(context, pointHTMatrix, color);
+    this.ticksPoints = ticksPoints;
 
+    context.resetTransform();
+    this.drawTicksTexts(ticksTexts, color, context);
     this.computeTextBoxes(context);
-    this.drawTitle(context, canvasHTMatrix);
+    this.drawTitle(context, canvasHTMatrix, color);
+
     context.setTransform(canvasHTMatrix);
     this.drawRubberBand(context);
   }
 
-  private drawTitle(context: CanvasRenderingContext2D, canvasHTMatrix: DOMMatrix) {
+  private drawTitle(context: CanvasRenderingContext2D, canvasHTMatrix: DOMMatrix, color: string): void {
     if (this.centeredTitle) {
       var [nameCoords, align, baseline, orientation] = this.centeredTitleProperties();
     } else {
@@ -2265,7 +2289,7 @@ export class newAxis {
     }
     nameCoords.transformSelf(canvasHTMatrix);
     const textParams: textParams = {
-      width: this.drawLength, fontsize: this.FONT_SIZE, font: this.FONT, align: align, color: this.strokeStyle,
+      width: this.drawLength, fontsize: this.FONT_SIZE, font: this.FONT, align: align, color: color,
       baseline: baseline, style: 'bold', orientation: orientation, backgroundColor: "hsla(0, 0%, 100%, 0.5)"
     };
     const textName = new newText(this.title, nameCoords, textParams);
@@ -2297,33 +2321,37 @@ export class newAxis {
     return [nameCoords, "center", baseline, this.isVertical ? -90 : 0]
   }
 
-  private drawTicks(context: CanvasRenderingContext2D, pointHTMatrix: DOMMatrix, color: string) {
-    const ticksCoords = [];
+  private drawTicksPoints(context: CanvasRenderingContext2D, pointHTMatrix: DOMMatrix, color: string): [newPoint2D[], newText[]] {
+    const ticksPoints = [];
     const ticksText: newText[] = [];
     const tickTextParams = this.computeTickTextParams();
 
     let count = Math.max(0, this.ticks[0]);
     this.ticks.forEach((tick, idx) => {
       if (tick >= this.minValue && tick <= this.maxValue) {
-        context.setTransform(pointHTMatrix);
         let point = this.drawTickPoint(context, tick, this.isVertical, pointHTMatrix, color);
         let text = this.labels[idx];
-        ticksCoords.push(point);
+        ticksPoints.push(point);
         if (this.isDiscrete) {
-          if (count == tick && this.labels[count]) { text = this.labels[count]; count++ }
+          if (count == tick && this.labels[count]) { text = this.labels[count] ; count++ }
           else { text = '' }
         }
         ticksText.push(this.computeTickText(context, text, tickTextParams, point, pointHTMatrix));
       }
     })
-    this.ticksFontsize = Math.min(...ticksText.map(tickText => tickText.fontsize));
-    context.resetTransform();
-    ticksText.forEach(tickText => {
-      tickText.fontsize = this.ticksFontsize;
-      tickText.width = null;
-      tickText.draw(context);
-    })
-    return ticksCoords
+    return [ticksPoints, ticksText]
+  }
+
+  private drawTicksTexts(ticksTexts: newText[], color: string, context: CanvasRenderingContext2D): void {
+    this.ticksFontsize = Math.min(...ticksTexts.map(tickText => tickText.fontsize));
+    ticksTexts.forEach(tickText => { this.drawTickText(tickText, color, context) });
+  }
+
+  private drawTickText(tickText: newText, color: string, context: CanvasRenderingContext2D): void {
+    tickText.fillStyle = color;
+    tickText.fontsize = this.ticksFontsize;
+    tickText.width = null;
+    tickText.draw(context);
   }
 
   private computeTickTextParams(): textParams {
@@ -2362,7 +2390,7 @@ export class newAxis {
     return tickText
   }
 
-  private getValueToDrawMatrix() {
+  private getValueToDrawMatrix(): DOMMatrix {
     const scale = this.drawLength / this.interval;
     return new DOMMatrix([
       scale, 0, 0, scale,
@@ -2377,7 +2405,7 @@ export class newAxis {
       maxValue * (1 + Math.sign(maxValue) * this.marginRatio)];
   }
 
-  public drawRubberBand(context: CanvasRenderingContext2D) {
+  public drawRubberBand(context: CanvasRenderingContext2D): void {
     const realMin = this.relativeToAbsolute(this.rubberBand.minValue);
     const realMax = this.relativeToAbsolute(this.rubberBand.maxValue);
     const coord = this.isVertical ? "y" : "x";
@@ -2388,7 +2416,7 @@ export class newAxis {
     this.rubberBand.draw(this.isVertical ? this.origin.x : this.origin.y, context, this.rubberColor, 'hsl(203, 0%, 100%, 0.5)', 0.1, 1.);
   }
 
-  public mouseMove(mouseDown: Vertex, mouseCoords: Vertex) {
+  public mouseMove(mouseDown: Vertex, mouseCoords: Vertex): boolean {
     let downValue = this.absoluteToRelative(this.isVertical ? mouseDown.y : mouseDown.x);
     let currentValue = this.absoluteToRelative(this.isVertical ? mouseCoords.y : mouseCoords.x);
     if (!this.rubberBand.isClicked) {
@@ -2398,7 +2426,7 @@ export class newAxis {
     return true
   }
 
-  public mouseDown(mouseDown: Vertex) {
+  public mouseDown(mouseDown: Vertex): boolean {
     let isReset = false;
     this.is_drawing_rubberband = true; // OLD
     const mouseUniCoord = this.isVertical ? mouseDown.y : mouseDown.x;
@@ -2409,7 +2437,7 @@ export class newAxis {
     return isReset
   }
 
-  public mouseUp() {
+  public mouseUp(): void {
     this.rubberBand.mouseUp();
     this.is_drawing_rubberband = false; // OLD
   }